refactor: 컴포넌트 폴더를 제거하여 depth 하나 낮춤
