+++ REQUIREMENTS.md
@@ -27,7 +27,7 @@
 - [ ] 불필요한 상수 파일 분리 하나로 합치기
 - [ ] 배포될 때 `data-cy` 제거하기
 - [ ] `RestaurantFilterContainer`를 `SearchBar`로 이름 변경하기
-- [ ] 컴포넌트 폴더 없애 depth 낮추기
+- [x] 컴포넌트 폴더 없애 depth 낮추기
 - [ ] `CategoryIcon` 컴포넌트 개선하기
   - 디자인을 구체적으로 결정하지 않기 (범용적으로 사용할 수 있게)
   - 도메인 의존성이 없는 방향으로 분리하기
