+++ src/components/CategoryIcon.styles.ts
