+++ src/components/CategoryIcon.tsx
@@ -1,12 +1,12 @@
-import type { Category } from '../../constants/categories';
+import type { Category } from '../constants/categories';
 import * as styled from './CategoryIcon.styles';
 
-import 기타 from '../../assets/categories/기타.png';
-import 아시안 from '../../assets/categories/아시안.png';
-import 양식 from '../../assets/categories/양식.png';
-import 일식 from '../../assets/categories/일식.png';
-import 중식 from '../../assets/categories/중식.png';
-import 한식 from '../../assets/categories/한식.png';
+import 기타 from '../assets/categories/기타.png';
+import 아시안 from '../assets/categories/아시안.png';
+import 양식 from '../assets/categories/양식.png';
+import 일식 from '../assets/categories/일식.png';
+import 중식 from '../assets/categories/중식.png';
+import 한식 from '../assets/categories/한식.png';
 
 const categoryToImageMapper: Record<Category, string> = {
   한식,
