+++ src/components/common/BottomSheet.styles.ts
